massive refactor for investment lot and transaction entry card
